Computer Vision projects/digit recognizer/requirements.txt
numpy=1.19.5
tensorflow=2.4.1
Keras=2.4.3
Pillow=7.2.0